Move configuration classes to own namespace
