+++ src/call_center/configuration/configuration_updater.cc
@@ -1,6 +1,6 @@
 #include "configuration_updater.h"
 
-namespace call_center {
+namespace call_center::config {
 
 std::shared_ptr<ConfigurationUpdater> ConfigurationUpdater::Create(
     std::shared_ptr<Configuration> configuration,
@@ -27,12 +27,18 @@ void ConfigurationUpdater::StartUpdating() {
   ScheduleUpdating();
 }
 
+void ConfigurationUpdater::AddUpdateListener(OnUpdate listener) {
+  listener(configuration_);
+  update_listeners_.emplace_back(std::move(listener));
+}
+
 void ConfigurationUpdater::ScheduleUpdating() {
   UpdateUpdatingPeriod();
   logger_->Info() << "Schedule configuration updating after " << updating_period_;
   task_manager_->PostTaskDelayed(updating_period_, [updater = shared_from_this()]() {
     updater->logger_->Info() << "Update configuration";
     updater->configuration_->UpdateConfiguration();
+    updater->NotifyListeners();
     updater->ScheduleUpdating();
   });
 }
@@ -43,4 +49,10 @@ void ConfigurationUpdater::UpdateUpdatingPeriod() {
       );
 }
 
-}  // namespace call_center
+void ConfigurationUpdater::NotifyListeners() const {
+  for (const auto &listener : update_listeners_) {
+    listener(configuration_);
+  }
+}
+
+}  // namespace call_center::config
