+++ src/call_center/core/tasks/task_manager_impl.cc
@@ -12,7 +12,7 @@ const size_t TaskManagerImpl::kDefaultIoThreadCount =
     std::max(static_cast<size_t>(boost::thread::hardware_concurrency()), static_cast<size_t>(64));
 
 std::shared_ptr<TaskManagerImpl> TaskManagerImpl::Create(
-    std::shared_ptr<Configuration> configuration, const log::LoggerProvider &logger_provider
+    std::shared_ptr<config::Configuration> configuration, const log::LoggerProvider &logger_provider
 ) {
   return std::shared_ptr<TaskManagerImpl>(
       new TaskManagerImpl(std::move(configuration), logger_provider)
@@ -20,7 +20,7 @@ std::shared_ptr<TaskManagerImpl> TaskManagerImpl::Create(
 }
 
 TaskManagerImpl::TaskManagerImpl(
-    std::shared_ptr<Configuration> configuration, const log::LoggerProvider &logger_provider
+    std::shared_ptr<config::Configuration> configuration, const log::LoggerProvider &logger_provider
 )
     : user_work_guard_(make_work_guard(user_context_)),
       io_work_guard_(make_work_guard(io_context_)),
